add G2 output option & bugfix in sedov

+++ src/globals.h
@@ -165,7 +165,7 @@ extern struct GasParticleData {
 
 extern float ( *Density_Func_Ptr ) ( const int, const double );
 extern float ( *U_Func_Ptr ) ( const int );
-extern void ( *Velocity_Func_Ptr ) ( const int, float * );
+extern void ( *Velocity_Func_Ptr ) ( const int, double * );
 extern void ( *Magnetic_Field_Func_Ptr ) ( const int, float * );
 extern void ( *PostProcessing_Func_Ptr ) ();
