+++ src/problems/blob.c
@@ -36,7 +36,7 @@ float Blob_Density ( const int ipart , const double bias )
 
 }
 
-void Blob_Velocity ( const int ipart, float out[3] )
+void Blob_Velocity ( const int ipart, double out[3] )
 {
     out[0] = 1000.0;
     out[1] = 0.0;
