+++ src/problems/evrard.c
@@ -35,7 +35,7 @@ float Evrard_Collapse_Density ( const int ipart , const double bias )
 
 }
 
-void Evrard_Collapse_Velocity ( const int ipart, float out[3] )
+void Evrard_Collapse_Velocity ( const int ipart, double out[3] )
 {
     out[0] = 0.0;
     out[1] = 0.0;
