+++ src/problems/galaxy_cluster.h
@@ -1,7 +1,7 @@
 float betamodel ( const float r );
 float GalaxyCluster_Density ( const int ipart , const double bias );
 
-void GalaxyCluster_Velocity ( const int ipart, float out[3] );
+void GalaxyCluster_Velocity ( const int ipart, double out[3] );
 float GalaxyCluster_U ( const int ipart );
 
 extern struct HaloProperties {
